* KaScPar does not like size_t replace by unsigned int
git-archimport-id: falk@cs.fau.de--2008-systemoc-intern/SysteMoC-Testcases--mainline--0.1--patch-50

+++ systemoc-fir/design-fir.cpp
@@ -65,7 +65,7 @@ private:
   }
   smoc_firing_state start;
 public:
-  m_h_src(sc_module_name name, size_t iterations)
+  m_h_src(sc_module_name name, unsigned int iterations)
     : smoc_actor(name, start),
       i(1) {
     start = out(1) >> (VAR(i) < iterations) >> call(&m_h_src::src) >> start;
@@ -156,7 +156,7 @@ public:
     return retval;
   }*/
   
-  m_h_top(sc_module_name name, size_t iterations)
+  m_h_top(sc_module_name name, unsigned int iterations)
     : smoc_graph(name),
       src("src", iterations),
       fir("fir"/*, gentaps()*/),
@@ -167,7 +167,7 @@ public:
 };
 
 int sc_main (int argc, char **argv) {
-  size_t iterations = 10000000; // ten million
+  unsigned int iterations = 10000000; // ten million
   if (argc >= 2)
     iterations = atoi(argv[1]);
   smoc_top_moc<m_h_top> top("top", iterations);
